tds: Avoid warning using poll() function
Due to some signed/not-signed calling poll with int type was
causing a warning under Linux.
All systems use unsigned number for length of poll

+++ src/replacements/poll.c
@@ -44,7 +44,7 @@
 #include <assert.h>
 
 int
-tds_poll(struct pollfd fds[], int nfds, int timeout)
+tds_poll(struct pollfd fds[], size_t nfds, int timeout)
 {
 	struct timeval tv, *tvp;
 	fd_set fdsr, fdsw, fdsp;
